backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.9
pydantic==2.10.4
python-dateutil==2.9.0.post0
ofxparse==0.21
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
pdfplumber==0.11.4
